src/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
click==8.1.7
coverage==7.4.3
Django==5.0
django-countries==7.5.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.0.0
flake8==7.0.0
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.26
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
yarl==1.9.4
